+++ modularcalculator/features/state/assignment.py
@@ -50,7 +50,7 @@ class AssignmentFeature(Feature):
 
         calculator.vars = {}
 
-        calculator.validators['variable'] = AssignmentFeature.validate_variable
+        calculator.add_validator('variable', 'variable', AssignmentFeature.validate_variable)
 
     var_regex = re.compile(r'([a-zA-Z_][a-zA-Z0-9_]*)')