+++ modularcalculator/numericalengine.py
@@ -17,7 +17,10 @@ class NumericalEngine(Engine):
         self.number_set_rounding('ROUND_HALF_UP')
 
         self.number_types = {}
-        self.validators['number'] = NumericalEngine.validate_number
+        self.numerical_validators = []
+        self.add_validator('number', 'number', NumericalEngine.validate_number, True)
+        self.add_validator('number_int', 'integer', NumericalEngine.validate_number_int, True)
+        self.add_validator('number_int_positive', 'positive integer', NumericalEngine.validate_number_int_positive, True)
 
     def number_size_set(self, size):
         self.number_size = size
@@ -39,6 +42,16 @@ class NumericalEngine(Engine):
     def number_set_rounding(self, rounding):
         Number.set_rounding(rounding)
 
+    def add_validator(self, name, desc, ref, is_number=False):
+        if name in self.validators:
+            raise Exception("Duplicate validator {}".format(name))
+        self.validators[name] = {
+            'desc': desc,
+            'ref': ref,
+        }
+        if is_number:
+            self.numerical_validators.append(name)
+
     def number(self, val):
         if isinstance(val, Number):
             return val
@@ -52,5 +65,19 @@ class NumericalEngine(Engine):
         try:
             self.number(value)
             return True
+        except Exception:
+            return False
+
+    def validate_number_int(self, value, unit=None, ref=None):
+        try:
+            num = self.number(value)
+            return (num % Number(1)).to_decimal() == 0
+        except Exception:
+            return False
+
+    def validate_number_int_positive(self, value, unit=None, ref=None):
+        try:
+            num = self.number(value)
+            return (num % Number(1)).to_decimal() == 0 and num.to_decimal() >= 0
         except Exception:
             return False